requirements.txt
django~=5.0.4
pillow
psycopg2
django-environ
python-decouple
django-filter
django-crispy-forms
django-admin-interface
django-mptt~=0.16.0
django-debug-toolbar
djangoql~=0.18.1
django-import-export
django-admin-colors
django-bootstrap-static
django-bootstrap5
bootstrap-admin
django-extensions
channels
redis
django-admin-search
django-jazzmin
django-admin-material-dashboard
django-redis